Created footer content
In the same way the Header was created, a footer partial was added.
The footer contains three elements.
A list of contacts. (six columns)
A footer menu. (six columns)
These are arranged with source ordering so the menu appears last on
mobile, but in the left column on larger screens.
The last element is a copyright that spans the full width of the page.